Bot/ccxt-trading-bot/requirements.txt
ccxt
pandas
pyyaml
python-dotenv
numpy
